feat: Refactoring for add objectId

+++ src/post/post.service.ts
@@ -5,6 +5,7 @@ import { Amuwiki } from 'src/amuwiki/schema/amuwiki.schema';
 import { AuthService } from 'src/auth/auth.service';
 import { User } from 'src/user/schemas/user.schema';
 import { CreatePostDto } from './dto/createpost.dto';
+import { DeletePostDto } from './dto/deletepost.dto';
 import { EditPostDto } from './dto/editpost.dto';
 
 @Injectable()
@@ -38,22 +39,50 @@ export class PostService {
     return await new this.amuwikiModel(temp).save();
   }
 
-  async editPost(email: string, dto: EditPostDto) {
-    const check = await this.amuwikiModel.findOne({ contributors: email });
+  async editPost(
+    email: string,
+    dto: EditPostDto,
+  ): Promise<{ message: string; statusCode: number }> {
+    const post = await this.amuwikiModel.findOne({ _id: dto._id });
 
+    if (!post) {
+      return { message: '게시글이 존재하지 않습니다.', statusCode: 400 };
+    }
+
+    if (email !== post.contributors[0]) {
+      return { message: '게시글의 소유자가 아닙니다.', statusCode: 400 };
+    }
     const temp = {
+      _id: dto._id,
       namespace: 0,
       title: dto.title,
       text: dto.text,
       contributors: email,
     };
-    Object.assign(check, temp);
-    return check.save();
+    Object.assign(post, temp);
+
+    post.save();
+    return { message: '게시글이 수정되었습니다.', statusCode: 201 };
   }
 
-  async deletePost(email: string) {
-    return await this.amuwikiModel.deleteOne({
-      contributors: email,
+  async deletePost(
+    email: string,
+    dto: DeletePostDto,
+  ): Promise<{ message: string; statusCode: number }> {
+    const post = await this.amuwikiModel.findOne({ _id: dto._id });
+
+    if (!post) {
+      return { message: '게시글이 존재하지 않습니다.', statusCode: 400 };
+    }
+
+    if (email !== post.contributors[0]) {
+      return { message: '게시글의 소유자가 아닙니다.', statusCode: 400 };
+    }
+
+    await this.amuwikiModel.deleteOne({
+      _id: dto._id,
     });
+
+    return { message: '게시글이 삭제되었습니다.', statusCode: 200 };
   }
 }
